feat: Windows compatibility
Changes include making the installer a python script (rather than a bash
script), and writing `internals/quik_setup.bat` to be used in Windows.

One critical change (resulting in a MAJOR version bump in
`internals/output_parse.py` is that ! (exclamation mark) is a reserved
character in Batch (when DelayedLocalExpansion is set, which it is).
This means that the mechanism of `quik.py` emitting

    !cd <new directory>

causes problems when being passed around from `quik.py` to
`internals/output_parse.py`. For this reason, the new character
signalling a command is `+`, such that `quik.py` emmits

    +cd <new directory>

I imagine this may cause trouble in the future, but we'll deal with that
when we get there.