fix: quik_setup.sh has been broken since 5e8dc402c
In 5e8dc402c3a320076cae0c376be2b980f6c92329, the commit where Windows
compatibility was introduced, the "flag" in the output to indicate a
change of directory went from `!cd` to `+cd` (because `!` is reserved in
batch). It seems like `quik_setup.sh` was never updated to account for
this change (which means quik has been broken since then).