fix: powershell script autocompletion was broken
When a user writes `quik <tab>`, PowerShell autocompleted this to `quik
quik`, because it made no distinction between this sequence of inputs
and `quik <tab>`. The expansion to `quik quik` was wrong, because doing
`quik <tab>` was being passed to the autocomplete python script as
`quik<tab>` (returning "quik", to mean leave the word as is). To fix
this, I just check if the cursor's position is greater than the
autocomplete string (which means there are spaces between the cursor and
the string to autocomplete, which is trimmed), and if so add a trailing
space, hinting the autocomplete python script that following words
should be suggested.